Only link bash_profile if it doesn't already exist.

+++ install.sh
@@ -1,5 +1,8 @@
 cwd=$(pwd)
 
-ln -s $cwd/.bash_profile $HOME
+if [ ! -e $cwd/.bash_profile ]; then
+  ln -s $cwd/.bash_profile $HOME
+fi
+
 ln -s $cwd/.vim $HOME
 ln -s $cwd/.vimrc $HOME
